requirements.txt
# using python 3 and requirements.txt
requests==2.19.1
google-api-python-client==1.7.4
oauth2client==4.1.2
